IOS and Android changes to build libindy, libnullpay, and libvcx (#373)
* Fixed the iOS build to link in the libnullpay library

* Now the iOS build links in the correct libssl libcrypto libnullpay libsodium libzmq and libindhy

* Added the rm statements back in and removed the lipo thinning

* Changed the method signature for importWallet to use the parameter name decryptionKey

* scripts for a rapid cocoapod dev cycle

* turned on the execute bit

* removed the sudo commands as they are no longer needed

* Added the chmod command so that the norm user can move the cocoapod to a different location

* Adding a python script to automatically instrument the source code with trace statements

* Now the logic works to add in trace statements if the following line ends in a semicolon

* Now the extern and use statments are ignored

* vcx/libvcx/build_scripts/add.trace.statements.to.src.py

* Adding more logic to automatically instrument the rust code with trace statements

* Adding more logic to automatically instrument the rust code with trace statements

* Now the automated instrumenting of the code works

* Making the script idempotent

* Changed the trace to println and it is now working

* Changed the trace message and added a new condition

* Adding logic to build a trace noclean version of libnullpay

* Added a few more conditions

* Adding more conditions to instrument the indy code base as well

* Finally the libindy code is instrumented to see if it invokes from vcx into libindy

* Fixed the script to really be idempotent

* Remove cargo update, Update IOS wrapper to fix credential flow, Fix Android build scripts

* Added the libnullpay dependencies to the vcx.framework build and to the library combining script

* Removed libnullpay.a dependency as it is not needed

+++ vcx/libvcx/build_scripts/ios/mac/mac.03.libindy.build.sh
@@ -10,6 +10,12 @@ SHA_HASH_DIR=$START_DIR/../..
 SHA_HASH_DIR=$(abspath "$SHA_HASH_DIR")
 
 source ./mac.02.libindy.env.sh
+
+CLEAN_BUILD="cleanbuild"
+if [ ! -z "$3" ]; then
+    CLEAN_BUILD=$3
+fi
+
 if [ -d $WORK_DIR/vcx-indy-sdk ]; then
     #rm -rf $WORK_DIR/vcx-indy-sdk
     cd $WORK_DIR/vcx-indy-sdk
@@ -17,14 +23,20 @@ else
     git clone https://github.com/hyperledger/indy-sdk.git $WORK_DIR/vcx-indy-sdk
     cd $WORK_DIR/vcx-indy-sdk
 fi
-git checkout .
-git checkout master
-git clean -f
-git clean -fd
-git pull
-git checkout `cat $SHA_HASH_DIR/libindy.commit.sha1.hash.txt`
-#cd $WORK_DIR/vcx-indy-sdk
-#git checkout tags/v1.3.0
+
+if [ "$CLEAN_BUILD" = "cleanbuild" ]; then
+    git checkout .
+    git checkout master
+    git clean -f
+    git clean -fd
+    git pull
+    git checkout `cat $SHA_HASH_DIR/libindy.commit.sha1.hash.txt`
+    #cd $WORK_DIR/vcx-indy-sdk
+    #git checkout tags/v1.3.0
+else
+    git checkout -- libindy/Cargo.toml
+    git checkout -- libnullpay/Cargo.toml
+fi
 
 DEBUG_SYMBOLS="debuginfo"
 if [ ! -z "$1" ]; then
@@ -45,12 +57,17 @@ if [ "$DEBUG_SYMBOLS" = "debuginfo" ]; then
     cat $START_DIR/cargo.toml.add.debug.txt >> Cargo.toml
 fi
 
-cargo clean
-cargo update
+if [ "$CLEAN_BUILD" = "cleanbuild" ]; then
+    cargo clean
+    # cargo update
+fi
+
 # To build for macos
 #cargo build
 # To build for iOS
+#echo "cargo lipo --release --verbose --targets=${IOS_TARGETS}"
 cargo lipo --release --verbose --targets="${IOS_TARGETS}"
+# cargo lipo --release --targets="${IOS_TARGETS}"
 #cargo lipo
 
 #########################################################################################################################
@@ -62,13 +79,18 @@ if [ "$DEBUG_SYMBOLS" = "debuginfo" ]; then
     cat $START_DIR/cargo.toml.add.debug.txt >> Cargo.toml
 fi
 
-cargo clean
-cargo update
+if [ "$CLEAN_BUILD" = "cleanbuild" ]; then
+    cargo clean
+    # cargo update
+fi
+
 # Replace '\"cdylib\"' with '\"staticlib\", \"cdylib\"' in Cargo.toml
 #sed -i .bak 's/\"cdylib\"/\"staticlib\", \"cdylib\"/' Cargo.toml
 
 # To build for macos
 #cargo build
 # To build for iOS
+#echo "cargo lipo --release --verbose --targets=${IOS_TARGETS}"
 cargo lipo --release --verbose --targets="${IOS_TARGETS}"
+# cargo lipo --release --targets="${IOS_TARGETS}"
 #cargo lipo
